requirements.txt
youtube_transcript-api
streamlit
google-generativeai
python-dotenv
pathlib
